audio-scam-detection/requirements.txt
autogen-agentchat==0.4.0.dev11
autogen-ext[openai]==0.4.0.dev11
pillow==11.0.0
pyyaml==6.0.2
python-dotenv==1.0.1
pymongo>=4.6.0
gradio>=5.0.0
openai-whisper
librosa
soundfile
requests>=2.31.0

# Email dependencies (Python fallback)
secure-smtplib>=0.1.1
